chore: split playwright.fixtures into files (#3982)
